2012-08-14 01:01 UTC - kode54 - Implemented VGM support for YM2203, YM3812, and YMF262
2012-08-13 09:46 UTC - kode54
- Implemented secondary resampler so VGM PCM chips are resampled to the FM mix rate
- Reset YM2413 volume level
- Implemented VGM support for OKIM6258, OKIM6295, K051649, K053260, and K054539

+++ gme/Ym2413_Emu.h
@@ -16,7 +16,7 @@ public:
 	
 	// Sets output sample rate and chip clock rates, in Hz. Returns non-zero
 	// if error.
-	int set_rate( double sample_rate, double clock_rate );
+	int set_rate( int sample_rate, int clock_rate );
 	
 	// Resets to power-up state
 	void reset();
